Update counter example
