--- examples/counter/index.js
@@ -1,7 +1,0 @@
-import React from 'react';
-import App from './containers/App';
-
-React.render(
-  <App />,
-  document.getElementById('root')
-);
