+++ examples/counter/package.json
@@ -16,16 +16,20 @@
   },
   "homepage": "https://github.com/gaearon/redux-devtools#readme",
   "dependencies": {
-    "react": "^0.13.3",
+    "react": "^0.14.0-rc1",
+    "react-dom": "^0.14.0-rc1",
     "react-redux": "^3.0.0",
     "redux": "^3.0.0",
+    "redux-devtools-log-monitor": "^1.0.0-alpha-1",
     "redux-thunk": "^1.0.0"
   },
   "devDependencies": {
     "babel-core": "^5.6.18",
     "babel-loader": "^5.1.4",
     "node-libs-browser": "^0.5.2",
+    "react-dock": "0.0.3",
     "react-hot-loader": "^1.3.0",
+    "redux-devtools-log-monitor": "^1.0.0-alpha",
     "webpack": "^1.9.11",
     "webpack-dev-server": "^1.9.0"
   }
