--- examples/counter/reducers/index.js
@@ -1,1 +1,0 @@
-export { default as counter } from './counter';
