+++ examples/counter/src/actions/CounterActions.js
