+++ examples/counter/src/components/Counter.js
