+++ examples/counter/src/constants/ActionTypes.js
