+++ examples/counter/src/containers/CounterApp.js
