+++ examples/counter/src/reducers/counter.js
