+++ examples/counter/src/store/configureStore.js
@@ -1,0 +1,5 @@
+if (process.env.NODE_ENV === 'development') {
+  module.exports = require('./configureStore.prod');
+} else {
+  module.exports = require('./configureStore.dev');
+}
